move validate_metric_mode from searcher to auto_estimator
